# web/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Robot Operation Center (ROC). Internal automation platform to distribute and execute robots across different subnets.",
    "version": "1.0.0",
    "title": "ROC (Robot Operation Center)",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
    "/project": {
      "get": {
        "operationId": "getProjects",
        "summary": "List all projects",
        "tags": [
          "project"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createProject",
        "summary": "Create new project",
        "tags": [
          "project"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Request body is invalid"
          }
        }
      }
    },
    "/project/{projectId}": {
      "get": {
        "operationId": "getProjectById",
        "summary": "Get Project by Id",
        "tags": [
          "project"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "description": "Id of project",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateProject",
        "summary": "Update Project",
        "tags": [
          "project"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "description": "Id of project",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "operationId": "deleteProject",
        "summary": "Delete Project",
        "tags": [
          "project"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "description": "Id of project",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/job": {
      "get": {
        "operationId": "getJobs",
        "summary": "List all jobs by project id",
        "tags": [
          "job"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "description": "Id of project",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          },
          {
            "in": "query",
            "name": "agentId",
            "description": "Filter jobs by agent",
            "required": false,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          },
          {
            "in": "query",
            "name": "environmentId",
            "description": "Filter jobs by environment",
            "required": false,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          },
          {
            "in": "query",
            "name": "taskForceId",
            "description": "Filter jobs by task force",
            "required": false,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "project object to be returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "triggerJob",
        "summary": "Execute new job",
        "tags": [
          "job"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a job object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          }
        }
      }
    },
...
    "/job/{jobId}/report": {
      "get": {
        "operationId": "getJobReport",
        "summary": "Get job report file",
        "tags": [
          "job"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "jobId",
            "description": "Identifier of job id",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "project object to be returned",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "AccessTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Access-Token",
        "description": "Provide access token to perform operations."
      }
    },
...
    "schemas": {
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          }
        }
      },
...
      "Job": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "taskForceId": {
            "type": "string"
          },
          "agentId": {
            "type": "string"
          },
          "environmentId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "IDLE",
              "RUNNING",
              "SUCCESS",
              "FAIL"
            ]
          }
        }
      }
    }
  }
}